+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 - Requires PHP `8.4`
 
+### Removed
+
+- `Innmind\HttpTransport\Curl::maxConcurrency()` use `::map()` instead
+
 ## 8.1.0 - 2025-09-18
 
 ### Added